move examples to src/ and convert followline and voicerecog to typescript
