+++ src/example/1.js
