+++ src/example/handRecognition/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <script src="cdn/tf.min.js"></script>
+        <script src="cdn/teachablemachine-image.min.js"></script>
+        <script src="cdn/socket.io.min.js"></script>
+
+        <title>Teachable Machine Image Model</title>
+    </head>
+    <div>Teachable Machine Image Model</div>
+
+    <button type="button" id="start-button">Start</button>
+    <div id="webcam-container"></div>
+    <div id="label-container"></div>
+
+    <script src="index.js" type="module"></script>
+</html>
