+++ src/example/handRecognition/model.json
