+++ src/example/handRecognition/weights.bin
