--- src/forceStop.js
@@ -1,7 +1,0 @@
-import { init, move } from '../out/lib.js';
-
-await init();
-
-// buzzer({pw:10, ms: 1000});
-
-move();
